backend/requirements.txt
censusgeocode==0.5.2
Django==4.0
djangorestframework==3.14.0
requests==2.28.1
